move alias to a HOC wrapper
